refactor: extract input monitors into AppDelegate+InputMonitors.swift, fix UserDefaults observer
